refactor: Cell を個別のフォルダに移動
