+++ src/components/game/cell/Cell.stories.tsx
@@ -3,7 +3,7 @@ import { ComponentStory, ComponentMeta } from '@storybook/react';
 
 import Cell from './Cell';
 import { within } from '@storybook/testing-library';
-import sleep from '../../utils/sleep';
+import sleep from '../../../utils/sleep';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
